+++ src/task3/task3.c
@@ -1,20 +1,20 @@
-#include "task3.h"
-#include "../utils/utils.h"
 #include <math.h>
 #include <mpi.h>
 #include <stdlib.h>
 #include <stdio.h>
+#include "task3.h"
+#include "../utils/utils.h"
 
 double f(double const x, double const y) {
     return 0.;
 }
 
 
-int64_t min(int64_t const a, int64_t const b) {
+int min(int const a, int const b) {
     return a < b ? a : b;
 }
 
-void TASK3_run(int64_t const n_points, double const eps, double const temperature, double const std, int const rank,
+void TASK3_run(int const n_points, double const eps, double const temperature, double const std, int const rank,
                int const comm_size) {
     double const h = 1. / ((double)n_points - 1);
     LAB2_matrix F, U;
@@ -22,8 +22,8 @@ void TASK3_run(int64_t const n_points, double const eps, double const temperatur
     matrix_alloc(n_points + 2, n_points + 2, &U);
 
     if (rank == 0) {
-        for (int64_t i = 0; i < n_points; ++i) {
-            for (int64_t j = 0; j < n_points; ++j) {
+        for (int i = 0; i < n_points; ++i) {
+            for (int j = 0; j < n_points; ++j) {
                 if (i == 0) {
                     *matrix_get(U, 0, j + 1) = *matrix_get(U, n_points + 1, j + 1) = temperature;
                 }
@@ -37,8 +37,8 @@ void TASK3_run(int64_t const n_points, double const eps, double const temperatur
             *matrix_get(U, n_points + 1, n_points + 1) = temperature;
     }
 
-    MPI_Bcast(U.array, (int)((n_points + 2) * (n_points + 2)), MPI_DOUBLE, 0, MPI_COMM_WORLD);
-    MPI_Bcast(F.array, (int)(n_points * n_points), MPI_DOUBLE, 0, MPI_COMM_WORLD);
+    MPI_Bcast(U.array, (n_points + 2) * (n_points + 2), MPI_DOUBLE, 0, MPI_COMM_WORLD);
+    MPI_Bcast(F.array, n_points * n_points, MPI_DOUBLE, 0, MPI_COMM_WORLD);
 
     double *dm = calloc(n_points, sizeof(double));
     double d_max = 1;
@@ -56,13 +56,13 @@ void TASK3_run(int64_t const n_points, double const eps, double const temperatur
             dm[i] = 0;
         }
 
-        for (int64_t n = 1; n < n_points + 1; ++n) {
-            int64_t const start = n / comm_size * rank + min(rank, n % comm_size),
-                          end = start + n / comm_size + (rank < n % comm_size ? 1 : 0);
-            int64_t counter = 0;
+        for (int n = 1; n < n_points + 1; ++n) {
+            int const start = n / comm_size * rank + min(rank, n % comm_size),
+                      end = start + n / comm_size + (rank < n % comm_size ? 1 : 0);
+            int counter = 0;
 
-            for (int64_t i = start + 1; i < end + 1; ++i) {
-                int64_t const j = n - i + 1;
+            for (int i = start + 1; i < end + 1; ++i) {
+                int const j = n - i + 1;
 
                 double const tmp = *matrix_get(U, i, j);
                 sendbuf[counter] = 0.25 * (*matrix_get(U, i - 1, j) + *matrix_get(U, i + 1, j) + *
@@ -75,28 +75,28 @@ void TASK3_run(int64_t const n_points, double const eps, double const temperatur
                 }
             }
 
-            for (int64_t i = 0; i < comm_size; ++i) {
-                recvcounts[i] = (int)n / comm_size + (i < n % comm_size ? 1 : 0);
+            for (int i = 0; i < comm_size; ++i) {
+                recvcounts[i] = n / comm_size + (i < n % comm_size ? 1 : 0);
             }
-            for (int64_t i = 1; i < comm_size; ++i) {
+            for (int i = 1; i < comm_size; ++i) {
                 displs[i] = displs[i - 1] + recvcounts[i - 1];
             }
 
-            MPI_Allgatherv(sendbuf, (int)counter, MPI_DOUBLE, recvbuf, recvcounts, displs,MPI_DOUBLE, MPI_COMM_WORLD);
+            MPI_Allgatherv(sendbuf, counter, MPI_DOUBLE, recvbuf, recvcounts, displs,MPI_DOUBLE, MPI_COMM_WORLD);
 
-            for (int64_t i = 1; i < n + 1; ++i) {
-                int64_t const j = n - i + 1;
+            for (int i = 1; i < n + 1; ++i) {
+                int const j = n - i + 1;
                 *matrix_get(U, i, j) = recvbuf[i - 1];
             }
         }
 
-        for (int64_t n = n_points - 1; n > 0; --n) {
-            int64_t const start = n / comm_size * rank + min(rank, n % comm_size),
-                          end = start + n / comm_size + (rank < n % comm_size ? 1 : 0);
-            int64_t counter = 0;
+        for (int n = n_points - 1; n > 0; --n) {
+            int const start = n / comm_size * rank + min(rank, n % comm_size),
+                      end = start + n / comm_size + (rank < n % comm_size ? 1 : 0);
+            int counter = 0;
 
-            for (int64_t i = n_points - n + 1 + start; i < n_points - n + 1 + end; ++i) {
-                int64_t const j = 2 * n_points + 1 - i - n;
+            for (int i = n_points - n + 1 + start; i < n_points - n + 1 + end; ++i) {
+                int const j = 2 * n_points + 1 - i - n;
 
                 double const tmp = *matrix_get(U, i, j);
                 sendbuf[counter] = 0.25 * (*matrix_get(U, i - 1, j) + *matrix_get(U, i + 1, j) + *
@@ -109,17 +109,17 @@ void TASK3_run(int64_t const n_points, double const eps, double const temperatur
                 }
             }
 
-            for (int64_t i = 0; i < comm_size; ++i) {
-                recvcounts[i] = (int)n / comm_size + (i < n % comm_size ? 1 : 0);
+            for (int i = 0; i < comm_size; ++i) {
+                recvcounts[i] = n / comm_size + (i < n % comm_size ? 1 : 0);
             }
-            for (int64_t i = 1; i < comm_size; ++i) {
+            for (int i = 1; i < comm_size; ++i) {
                 displs[i] = displs[i - 1] + recvcounts[i - 1];
             }
 
-            MPI_Allgatherv(sendbuf, (int)counter, MPI_DOUBLE, recvbuf, recvcounts, displs,MPI_DOUBLE, MPI_COMM_WORLD);
+            MPI_Allgatherv(sendbuf, counter, MPI_DOUBLE, recvbuf, recvcounts, displs,MPI_DOUBLE, MPI_COMM_WORLD);
 
-            for (int64_t i = n_points - n + 1; i < n_points + 1; ++i) {
-                int64_t const j = 2 * n_points + 1 - i - n;
+            for (int i = n_points - n + 1; i < n_points + 1; ++i) {
+                int const j = 2 * n_points + 1 - i - n;
                 *matrix_get(U, i, j) = recvbuf[i - n_points + n - 1];
             }
         }
@@ -136,8 +136,8 @@ void TASK3_run(int64_t const n_points, double const eps, double const temperatur
     if (rank == 0) {
         FILE *file = fopen("mat.txt", "w+");
 
-        for (int64_t i = 0; i < n_points; ++i) {
-            for (int64_t j = 0; j < n_points; ++j) {
+        for (int i = 0; i < n_points; ++i) {
+            for (int j = 0; j < n_points; ++j) {
                 fprintf(file, "%e ", *matrix_get(U, i + 1, j + 1));
             }
             fprintf(file, "\n");
