+++ src/task3/task3.h
@@ -1,8 +1,6 @@
 #ifndef TASK3_H
 #define TASK3_H
 
-#include <stdlib.h>
-
-void TASK3_run(int64_t n_points, double eps, double temperature, double std, int rank, int comm_size);
+void TASK3_run(int n_points, double eps, double temperature, double std, int rank, int comm_size);
 
 #endif
